Use byte instead of int in the lexer

+++ src/main/java/minijava/lexer/LexerInput.java
@@ -2,15 +2,15 @@ package minijava.lexer;
 
 import java.util.Iterator;
 
-/** Input for a lexer (a stream of integers) it stops after the EOF character is read. */
-interface LexerInput extends Iterator<Integer> {
+/** Input for a lexer (a stream of bytes) it stops after the EOF character is read. */
+interface LexerInput extends Iterator<Byte> {
 
   /** Close the underlying streams and clean up if needed. */
   void close();
 
   Location getCurrentLocation();
 
-  int current();
+  byte current();
 
   default boolean isCurrentChar(char... eitherChars) {
     for (char eitherChar : eitherChars) {
